+++ src/App/components/CategoriesList.module.scss
