+++ src/App/components/SearchInput.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { SetURLSearchParams, TQueryParams } from '../../types/types';
+import { SetURLSearchParams, TQueryParams } from '../types/types';
 
 type TProps = {
   query: string;
