+++ src/App/components/layouts/MainLayout.tsx
@@ -1,6 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import { Header } from '../Header/Header';
-// import styles from './styles.module.scss';
+import { Header } from '../Header';
 
 function Layout() {
   return (
